seperated out tests to test folder
