+++ src/server/controllers/axios.helper.ts
@@ -1,21 +1,8 @@
-import axios, { AxiosError, AxiosResponse } from 'axios';
+import axios, { AxiosError } from 'axios';
 
-export function errorNotFound(error: unknown | AxiosError): boolean {
+export default function errorNotFound(error: unknown | AxiosError): boolean {
   if (axios.isAxiosError(error)) {
     return error.response?.status === 404;
   }
   return false;
-}
-
-export function createAxiosError(responceStatus: number) {
-  const axiosResponse: AxiosResponse = {
-    status: responceStatus,
-  } as AxiosResponse;
-
-  return {
-    config: {},
-    request: {},
-    response: axiosResponse,
-    isAxiosError: true,
-  } as unknown as AxiosError<unknown>;
 }
