+++ src/server/controllers/case.controller.ts
@@ -4,7 +4,7 @@ import { IControllerInterface } from '../interfaces/controller.interface';
 import { IConfiguration } from '../interfaces/configuration.interface';
 import { ICaseDetails } from '../interfaces/case.details.interface';
 import mapCaseDetails from '../mappers/case.details.mapper';
-import { errorNotFound } from './axios.helper';
+import errorNotFound from './axios.helper';
 
 export default class CaseController implements IControllerInterface {
   config: IConfiguration;
@@ -25,7 +25,7 @@ export default class CaseController implements IControllerInterface {
   async getListOfCases(request: Request, response: Response<ICaseDetails[]>) {
     const { questionnaireName } = request.params;
 
-    if (typeof questionnaireName !== 'string') {
+    if (questionnaireName === undefined) {
       throw new Error('Questionnaire name has not been provided');
     }
