+++ src/test/client/App.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 import { BrowserRouter } from 'react-router-dom';
-import App from './App';
+import App from '../../client/App';
 
 test('renders home page', () => {
   render(<BrowserRouter><App /></BrowserRouter>);
