+++ src/test/server/controllers/questionnaire.controller.test.ts
@@ -1,9 +1,9 @@
 import supertest, { Response } from 'supertest';
 import BlaiseClient, { IQuestionnaire, QuestionnaireListMockObject } from 'blaise-api-node-client';
 import { IMock, Mock, Times } from 'typemoq';
-import nodeServer from '../server';
+import nodeServer from '../../../server/server';
 import FakeConfiguration from '../configuration/configuration.fake';
-import { createAxiosError } from './axios.helper';
+import createAxiosError from './axios.test.helper';
 
 // create fake config
 const configFake = new FakeConfiguration('restapi.blaise.com', 'dist', 5000, 'gusty', 'cati.blaise.com');
